--- a/odl/contrib/tensorflow/layer.py
+++ b/odl/contrib/tensorflow/layer.py
@@ -11,8 +11,6 @@
 # Imports for common Python 2/3 codebase
 from __future__ import print_function, division, absolute_import
 from builtins import range, str
-from future import standard_library
-standard_library.install_aliases()
 
 import numpy as np
 import uuid