--- a/odl/contrib/theano/layer.py
+++ b/odl/contrib/theano/layer.py
@@ -10,9 +10,7 @@
 
 # Imports for common Python 2/3 codebase
 from __future__ import print_function, division, absolute_import
-from future import standard_library
 from future.utils import native
-standard_library.install_aliases()
 
 import theano
 import numpy as np