--- a/odl/diagnostics/space.py
+++ b/odl/diagnostics/space.py
@@ -10,8 +10,6 @@
 
 # Imports for common Python 2/3 codebase
 from __future__ import print_function, division, absolute_import
-from future import standard_library
-standard_library.install_aliases()
 from builtins import object
 
 from copy import copy, deepcopy