--- a/odl/discr/discr_mappings.py
+++ b/odl/discr/discr_mappings.py
@@ -14,8 +14,6 @@ operators.
 
 # Imports for common Python 2/3 codebase
 from __future__ import print_function, division, absolute_import
-from future import standard_library
-standard_library.install_aliases()
 from builtins import str, super, zip
 
 from itertools import product