--- a/odl/discr/lp_discr.py
+++ b/odl/discr/lp_discr.py
@@ -10,8 +10,6 @@
 
 # Imports for common Python 2/3 codebase
 from __future__ import print_function, division, absolute_import
-from future import standard_library
-standard_library.install_aliases()
 from builtins import super, str
 
 import numpy as np