--- a/odl/discr/partition.py
+++ b/odl/discr/partition.py
@@ -16,8 +16,6 @@ of partitions of intervals.
 
 # Imports for common Python 2/3 codebase
 from __future__ import print_function, division, absolute_import
-from future import standard_library
-standard_library.install_aliases()
 from builtins import object, range, super, zip
 
 import numpy as np