--- a/odl/operator/default_ops.py
+++ b/odl/operator/default_ops.py
@@ -10,11 +10,8 @@
 
 # Imports for common Python 2/3 codebase
 from __future__ import print_function, division, absolute_import
-from future import standard_library
-standard_library.install_aliases()
 from builtins import super
 
-import numpy as np
 from copy import copy
 import numpy as np