--- a/odl/phantom/transmission.py
+++ b/odl/phantom/transmission.py
@@ -10,8 +10,6 @@
 
 # Imports for common Python 2/3 codebase
 from __future__ import print_function, division, absolute_import
-from future import standard_library
-standard_library.install_aliases()
 
 from odl.discr import DiscreteLp
 from odl.phantom.geometric import ellipsoid_phantom