--- a/odl/set/space.py
+++ b/odl/set/space.py
@@ -10,9 +10,7 @@
 
 # Imports for common Python 2/3 codebase
 from __future__ import print_function, division, absolute_import
-from builtins import object, range, str
-from future import standard_library
-standard_library.install_aliases()
+from builtins import object, range
 
 import numpy as np