--- a/odl/solvers/functional/functional.py
+++ b/odl/solvers/functional/functional.py
@@ -8,8 +8,6 @@
 
 # Imports for common Python 2/3 codebase
 from __future__ import print_function, division, absolute_import
-from future import standard_library
-standard_library.install_aliases()
 from builtins import super
 
 import numpy as np