--- a/odl/solvers/nonsmooth/chambolle_pock.py
+++ b/odl/solvers/nonsmooth/chambolle_pock.py
@@ -14,8 +14,6 @@ non-smooth convex optimization problems in imaging.
 
 # Imports for common Python 2/3 codebase
 from __future__ import print_function, division, absolute_import
-from future import standard_library
-standard_library.install_aliases()
 
 import numpy as np