--- a/odl/solvers/smooth/nonlinear_cg.py
+++ b/odl/solvers/smooth/nonlinear_cg.py
@@ -10,9 +10,6 @@
 
 # Imports for common Python 2/3 codebase
 from __future__ import print_function, division, absolute_import
-from future import standard_library
-standard_library.install_aliases()
-
 
 from odl.solvers.util import ConstantLineSearch