--- a/odl/space/pspace.py
+++ b/odl/space/pspace.py
@@ -11,8 +11,6 @@
 # Imports for common Python 2/3 codebase
 from __future__ import print_function, division, absolute_import
 from builtins import range, str, super, zip
-from future import standard_library
-standard_library.install_aliases()
 
 from numbers import Integral
 from itertools import product