I did some import profiling and identified the main culprits regarding `import odl`. I write the list for reference, to avoid costly imports in the future:

- `scipy.linalg`, `scipy.sparse` and `scipy.special`: each one no large contribution, but together maybe 100 ms.
  Solution: lazy import. Usage is rare anyway.
- `skimage.transform`: takes almost 200 ms.
  Solution: `import skimage` at top level to set `SKIMAGE_AVAILABLE`,  import `radon, iradon` lazily. Big speedup.
- `future.standard_library.install_aliases()`: Does too much stuff, takes 50 to 100 ms. And adds annoying warnings in Spyder that shadow real warnings (e.g. unused imports).
  Solution: removed
- `pkg_resources`: Surprisingly, this one takes over 200 ms!
  Solution: I changed the version checks at module level to a simple homebrew solution and import lazily otherwise. 

The whole set of changes brings `import odl` down to 250 to 350 ms, which can almost be called snappy 🐊 .